(PC-42337) refactor(location): add tests and rename files
