+++ src/App.tsx
@@ -25,8 +25,8 @@ import { initAlgoliaAnalytics } from 'libs/algolia/analytics/initAlgoliaAnalytic
 import { env } from 'libs/environment/env'
 import { AnalyticsInitializer } from 'libs/firebase/analytics/AnalyticsInitializer'
 import { FirestoreNetworkObserver } from 'libs/firebase/firestore/FirestoreNetworkObserver/FirestoreNetworkObserver'
-import { GeolocationActivationModal } from 'libs/location/geolocation/components/GeolocationActivationModal'
-import { initLocation } from 'libs/locationV2/location.methods'
+import { GeolocationActivationModal } from 'libs/location/components/GeolocationActivationModal'
+import { initLocation } from 'libs/locationV2/initLocation'
 import { eventMonitoring } from 'libs/monitoring/services'
 import { NetInfoWrapper } from 'libs/network/NetInfoWrapper'
 import { OfflineModeContainer } from 'libs/network/OfflineModeContainer'
