+++ src/features/favorites/pages/FavoritesSorts.tsx
@@ -12,7 +12,7 @@ import { getTabHookConfig } from 'features/navigation/TabBar/getTabHookConfig'
 import { useGoBack } from 'features/navigation/useGoBack'
 import { analytics } from 'libs/analytics/provider'
 import { GeolocPermissionState, useLocation } from 'libs/location/location'
-import { requestGeolocPermission } from 'libs/locationV2/location.methods'
+import { requestGeolocPermission } from 'libs/locationV2/requestGeolocPermission'
 import { Button } from 'ui/designSystem/Button/Button'
 import { RadioButtonGroup } from 'ui/designSystem/RadioButtonGroup/RadioButtonGroup'
 import { PageWithHeader } from 'ui/pages/PageWithHeader'
