+++ src/features/onboarding/pages/onboarding/OnboardingGeolocation.tsx
@@ -4,7 +4,7 @@ import React, { useCallback } from 'react'
 import { getOnboardingHookConfig } from 'features/navigation/navigators/OnboardingStackNavigator/getOnboardingHookConfig'
 import { UseNavigationType } from 'features/navigation/navigators/RootNavigator/types'
 import { analytics } from 'libs/analytics/provider'
-import { requestGeolocPermission } from 'libs/locationV2/location.methods'
+import { requestGeolocPermission } from 'libs/locationV2/requestGeolocPermission'
 import { useMobileFontScaleToDisplay } from 'shared/accessibility/helpers/zoomHelpers'
 import Geolocation from 'ui/animations/geolocalisation.json'
 import { GenericInfoPage } from 'ui/pages/GenericInfoPage'
