+++ src/shared/Banners/GeolocationBanner.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent, useCallback } from 'react'
 import styled from 'styled-components/native'
 
 import { GeolocPermissionState, useLocation } from 'libs/location/location'
-import { requestGeolocPermission } from 'libs/locationV2/location.methods'
+import { requestGeolocPermission } from 'libs/locationV2/requestGeolocPermission'
 import { SystemBanner } from 'ui/components/ModuleBanner/SystemBanner'
 import { Everywhere } from 'ui/svg/icons/Everywhere'
